# CA-ARCH/api/openapi.yaml
openapi: 3.1.0
info:
  title: MarketHub Product REST API
  version: 1.0.0
paths:
  /products:
    get:
      summary: List products
      parameters:
        - {in: query, name: category, schema: {type: string}}
        - {in: query, name: status, schema: {type: string}}
        - {in: query, name: seller_id, schema: {type: string}}
      responses:
        '200': {description: Product list}
    post:
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ProductInput'}
      responses:
        '201': {description: Product created}
  /products/{product_id}:
    parameters:
      - {in: path, name: product_id, required: true, schema: {type: string}}
    get:
      summary: Get product
      responses:
        '200': {description: Product}
        '404': {description: Not found}
    patch:
      summary: Update product
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ProductPatch'}
      responses:
        '200': {description: Product updated}
        '404': {description: Not found}
    delete:
      summary: Delete product
      responses:
        '204': {description: Product deleted}
        '404': {description: Not found}
components:
  schemas:
    ProductInput:
      type: object
      required: [sku, title, price, seller_id]
      properties:
        sku: {type: string, minLength: 1}
        title: {type: string, minLength: 1}
        description: {type: string, default: ''}
        category: {type: string, default: general}
        price: {type: number, minimum: 0}
        currency: {type: string, default: RUB}
        stock_qty: {type: integer, minimum: 0, default: 0}
        status: {type: string, default: active}
        seller_id: {type: string, minLength: 1}
    ProductPatch:
      type: object
      properties:
        sku: {type: string, minLength: 1}
        title: {type: string, minLength: 1}
        description: {type: string}
        category: {type: string}
        price: {type: number, minimum: 0}
        currency: {type: string}
        stock_qty: {type: integer, minimum: 0}
        status: {type: string}
        seller_id: {type: string, minLength: 1}